feature test green
lots to refactor, and I'm not technically testing for the precense
of elements inside of the page, besides making sure that there are ten
and the right text count appears, but this is good enough for now.
I'll hit the model test and services next, to clean out the controller.